+++ src/setGlobalVars.js
@@ -69,6 +69,9 @@ function setGlobalVars (idb) {
         IDB.shimIndexedDB.__debug = function (val) {
             CFG.DEBUG = val;
         };
+        IDB.shimIndexedDB.__setCFG = function (prop, val) {
+            CFG[prop] = val;
+        };
     }
 
     // Workaround to prevent an error in Firefox